+++ src/BaseResponse.ts
@@ -1,6 +1,6 @@
 import { APIGatewayProxyEvent } from 'aws-lambda';
 import * as jwt from 'jsonwebtoken';
-import { IOptions, IOptionsDefined, Options, StatusCodeType } from './utils';
+import { IOptions, Options, StatusCodeType } from './utils';
 
 
 export interface IResponse {
@@ -13,23 +13,22 @@ export abstract class Response implements IResponse {
     readonly statusCode: number;
     readonly body: any;
     readonly headers: {[key: string]: any};
-    private readonly options: IOptionsDefined;
 
-    constructor(event: APIGatewayProxyEvent, statusCode: StatusCodeType, body?: any, options?: IOptions) {
+    constructor(event: APIGatewayProxyEvent, statusCode: StatusCodeType, body?: any, providedOptions?: IOptions) {
         this.statusCode = statusCode;
         this.body = JSON.stringify(body);
-        this.options = new Options(options);
-        this.headers = this.options.headers;
+        const options = new Options(providedOptions);
+        this.headers = options.headers;
 
-        this.logRequest(event);
+        this.logRequest(event, options);
     }
 
-    private logRequest(event: APIGatewayProxyEvent) {
-        const token = event.headers[this.options.tokenHeaderKey];
+    private logRequest(event: APIGatewayProxyEvent, options: Options) {
+        const token = event.headers[options.tokenHeaderKey];
 
-        this.options.logger.log(
+        options.logger.log(
             JSON.stringify({
-                user: token ? this.getUser(token) : undefined,
+                user: token ? this.getUser(token, options) : undefined,
                 responseStatus: this.statusCode,
                 timestamp: new Date(event.requestContext.requestTimeEpoch),
                 method: event.httpMethod,
@@ -37,19 +36,19 @@ export abstract class Response implements IResponse {
                 pathParameters: event.pathParameters,
                 query: event.queryStringParameters,
                 sourceIP: event.requestContext.identity.sourceIp,
-                body: this.options.hideBody ? '********' : event.body
+                body: options.hideBody ? '********' : event.body
             })
         );
     }
 
-    private getUser(token: string): string {
+    private getUser(token: string, options: Options): string {
         const decodedToken = jwt.decode(token);
-        const user = (decodedToken as jwt.JwtPayload)[this.options.tokenUserKey];
+        const user = (decodedToken as jwt.JwtPayload)[options.tokenUserKey];
 
         if (user) {
             return user;
         } else {
-            throw new Error(`JWT invalid. Does not contain property ${this.options.tokenUserKey}`);
+            throw new Error(`JWT invalid. Does not contain property ${options.tokenUserKey}`);
         }
     }
 }